app/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiolimiter==1.2.1
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
asgiref==3.5.0
async-timeout==5.0.1
attrs==25.1.0
autopep8==1.6.0
beautifulsoup4==4.12.3
billiard==4.2.1
black==25.1.0
boto3==1.24.67
botocore==1.27.67
bs4==0.0.2
cairocffi==1.7.1
CairoSVG==2.7.1
celery==5.4.0
certifi==2025.1.31
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.2.1
cron-descriptor==1.4.5
cryptography==41.0.3
cssselect2==0.8.0
cycler==0.12.1
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.8
distro==1.9.0
Django==4.0.1
django-celery-beat==2.7.0
django-cors-headers==4.2.0
django-ses==4.1.1
django-storages==1.13.1
django-templated-mail==1.1.1
django-timezone-field==7.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
djoser==2.2.0
exceptiongroup==1.2.2
filelock==3.13.1
fonttools==4.53.1
frozenlist==1.5.0
fsspec==2025.2.0
greenlet==3.1.1
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.29.1
idna==3.4
imageio==2.35.1
imageio-ffmpeg==0.5.1
imgkit==1.2.3
importlib_metadata==7.1.0
importlib_resources==6.4.3
Jinja2==3.1.5
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.5
kombu==5.4.2
Markdown==3.6
MarkupSafe==3.0.2
matplotlib==3.9.2
moviepy==1.0.3
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.2.1
numpy==2.0.1
oauthlib==3.2.2
openai==1.40.6
outcome==1.3.0.post0
packaging==23.1
pandas==2.2.2
pathspec==0.12.1
Pillow==9.1.0
pipenv==2023.5.19
platformdirs==4.2.0
playwright==1.51.0
proglog==0.1.10
prompt_toolkit==3.0.50
propcache==0.2.1
psycopg2==2.9.7
pycodestyle==2.8.0
pycparser==2.21
pydantic==2.8.2
pydantic_core==2.20.1
pyee==12.1.1
PyJWT==2.4.0
pyparsing==3.1.2
PySocks==1.7.1
python-crontab==3.2.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
requests==2.31.0
requests-oauthlib==1.3.1
resend==2.6.0
s3transfer==0.6.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
selenium==4.29.0
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.4.2
stripe==10.7.0
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tinycss2==1.4.0
tokenizers==0.21.0
toml==0.10.2
tomli==2.2.1
tqdm==4.66.5
transformers==4.49.0
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.12
vine==5.1.0
virtualenv==20.25.0
virtualenv-clone==0.5.7
voyageai==0.3.2
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==10.4
whitenoise==6.1.0
wsproto==1.2.0
yarl==1.18.3
zipp==3.18.1
